pyproject.toml
[project]
name = "business-assistant-v2"
version = "0.1.0"
description = "Plugin-based XMPP chatbot using PydanticAI for natural language understanding"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic-ai",
    "python-dotenv",
    "bot-commander[xmpp]",
    "business-assistant-imap-plugin",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/business_assistant"]

[tool.uv.sources]
bot-commander = { path = "D:/GIT/BenjaminKobjolke/bot-commander", editable = true }
business-assistant-imap-plugin = { path = "D:/GIT/BenjaminKobjolke/business-assistant-imap-plugin", editable = true }

[project.scripts]
business-assistant = "business_assistant.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
